Use (int) instead of (uint) for FFI
Our manual conversions do the same bitcast anyway, so we might as well change the FFI definitions instead for cleaner code.

+++ crates/bindings-csharp/Runtime/Internal/IReducer.cs
@@ -26,9 +26,9 @@ public interface IReducer
 
         FFI.volatile_nonatomic_schedule_immediate(
             name_bytes,
-            (uint)name_bytes.Length,
+            name_bytes.Length,
             args_bytes,
-            (uint)args_bytes.Length
+            args_bytes.Length
         );
     }
 }
