+++ crates/bindings-csharp/Runtime/Internal/ITable.cs
@@ -19,7 +19,7 @@ internal abstract class RawTableIterBase<T> : IEnumerable<T>
         {
             while (handle != FFI.RowIter.INVALID)
             {
-                var buffer_len = (uint)buffer.Length;
+                var buffer_len = buffer.Length;
                 var ret = FFI.row_iter_bsatn_advance(handle, buffer, ref buffer_len);
                 // On success, the only way `buffer_len == 0` is for the iterator to be exhausted.
                 // This happens when the host iterator was empty from the start.
@@ -34,7 +34,7 @@ internal abstract class RawTableIterBase<T> : IEnumerable<T>
                     // We got a chunk of rows, parse all of them before moving to the next chunk.
                     case Errno.OK:
                     {
-                        using var stream = new MemoryStream(buffer, 0, (int)buffer_len);
+                        using var stream = new MemoryStream(buffer, 0, buffer_len);
                         using var reader = new BinaryReader(stream);
                         while (stream.Position < stream.Length)
                         {
@@ -47,7 +47,7 @@ internal abstract class RawTableIterBase<T> : IEnumerable<T>
                     // The `buffer_len` will have been updated with the necessary size.
                     case Errno.BUFFER_TOO_SMALL:
                         ArrayPool<byte>.Shared.Return(buffer);
-                        buffer = ArrayPool<byte>.Shared.Rent((int)buffer_len);
+                        buffer = ArrayPool<byte>.Shared.Rent(buffer_len);
                         break;
                     default:
                         ret.Check();
@@ -92,7 +92,7 @@ public interface ITableView<View, T>
         new(() =>
         {
             var name_bytes = System.Text.Encoding.UTF8.GetBytes(tableName);
-            FFI.table_id_from_name(name_bytes, (uint)name_bytes.Length, out var out_);
+            FFI.table_id_from_name(name_bytes, name_bytes.Length, out var out_);
             return out_;
         });
 
@@ -118,11 +118,11 @@ public interface ITableView<View, T>
     {
         // Insert the row.
         var bytes = IStructuralReadWrite.ToBytes(row);
-        var bytes_len = (uint)bytes.Length;
+        var bytes_len = bytes.Length;
         FFI.datastore_insert_bsatn(tableId, bytes, ref bytes_len);
 
         // Write back any generated column values.
-        using var stream = new MemoryStream(bytes, 0, (int)bytes_len);
+        using var stream = new MemoryStream(bytes, 0, bytes_len);
         using var reader = new BinaryReader(stream);
         return View.ReadGenFields(reader, row);
     }
@@ -130,7 +130,7 @@ public interface ITableView<View, T>
     protected static bool DoDelete(T row)
     {
         var bytes = IStructuralReadWrite.ToBytes(row);
-        FFI.datastore_delete_all_by_eq_bsatn(tableId, bytes, (uint)bytes.Length, out var out_);
+        FFI.datastore_delete_all_by_eq_bsatn(tableId, bytes, bytes.Length, out var out_);
         return out_ > 0;
     }
